Remove all datanode leader node usages
