+++ data-node/src/main/java/org/graylog/datanode/bindings/PeriodicalBindings.java
@@ -19,7 +19,6 @@ package org.graylog.datanode.bindings;
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.Multibinder;
 import org.graylog.datanode.bootstrap.preflight.DataNodeConfigurationPeriodical;
-import org.graylog.datanode.periodicals.ClusterManagerDiscovery;
 import org.graylog.datanode.periodicals.MetricsCollector;
 import org.graylog.datanode.periodicals.NodePingPeriodical;
 import org.graylog.datanode.periodicals.OpensearchNodeHeartbeat;
@@ -34,7 +33,6 @@ public class PeriodicalBindings extends AbstractModule {
         periodicalBinder.addBinding().to(ClusterEventPeriodical.class);
         periodicalBinder.addBinding().to(ClusterEventCleanupPeriodical.class);
         periodicalBinder.addBinding().to(OpensearchNodeHeartbeat.class);
-        periodicalBinder.addBinding().to(ClusterManagerDiscovery.class);
 //        periodicalBinder.addBinding().to(UserSessionTerminationPeriodical.class);
         periodicalBinder.addBinding().to(NodePingPeriodical.class);
         periodicalBinder.addBinding().to(DataNodeConfigurationPeriodical.class);
