+++ data-node/src/main/java/org/graylog/datanode/opensearch/OpensearchProcess.java
@@ -34,9 +34,6 @@ public interface OpensearchProcess extends ManagableProcess<OpensearchConfigurat
     Optional<RestHighLevelClient> restClient();
 
     Optional<OpenSearchClient> openSearchClient();
-
-    boolean isLeaderNode();
-    void setLeaderNode(boolean isManagerNode);
     List<String> stdOutLogs();
     List<String> stdErrLogs();
