+++ data-node/src/test/java/org/graylog/datanode/periodicals/NodePingPeriodicalTest.java
@@ -30,7 +30,7 @@ import java.net.URI;
 class NodePingPeriodicalTest {
 
     @Test
-    void doRun() throws NodeNotFoundException {
+    void doRun() {
 
         final SimpleNodeId nodeID = new SimpleNodeId("5ca1ab1e-0000-4000-a000-000000000000");
         final URI uri = URI.create("http://localhost:9200");
@@ -49,7 +49,6 @@ class NodePingPeriodicalTest {
                 () -> uri,
                 () -> cluster,
                 () -> datanodeRestApi,
-                () -> true,
                 () -> OpensearchState.AVAILABLE
         );
